refactor(collapse): make collapse-panel a separate directory
