+++ example/app.json
@@ -99,56 +99,7 @@
     }
   ],
   "usingComponents": {
-    "t-demo": "./components/demo-block/index",
-    "t-avatar": "tdesign-miniprogram/avatar/avatar",
-    "t-avatar-group": "tdesign-miniprogram/avatar/avatar-group",
-    "t-badge": "tdesign-miniprogram/badge/badge",
-    "t-button": "tdesign-miniprogram/button/button",
-    "t-cell": "tdesign-miniprogram/cell/cell",
-    "t-cell-group": "tdesign-miniprogram/cell-group/cell-group",
-    "t-collapse": "tdesign-miniprogram/collapse/collapse",
-    "t-collapse-panel": "tdesign-miniprogram/collapse/collapse-panel",
-    "t-tabs": "tdesign-miniprogram/tabs/tabs",
-    "t-tab-panel": "tdesign-miniprogram/tabs/tab-panel",
-    "t-loading": "tdesign-miniprogram/loading/loading",
-    "t-icon": "tdesign-miniprogram/icon/icon",
-    "t-rate": "tdesign-miniprogram/rate/rate",
-    "t-progress": "tdesign-miniprogram/progress/progress",
-    "t-input": "tdesign-miniprogram/input/input",
-    "t-textarea": "tdesign-miniprogram/textarea/textarea",
-    "t-message": "tdesign-miniprogram/message/message",
-    "t-toast": "tdesign-miniprogram/toast/toast",
-    "t-stepper": "tdesign-miniprogram/stepper/stepper",
-    "t-slider": "tdesign-miniprogram/slider/slider",
-    "t-radio": "tdesign-miniprogram/radio/radio",
-    "t-radio-group": "tdesign-miniprogram/radio-group/radio-group",
-    "t-checkbox": "tdesign-miniprogram/checkbox/checkbox",
-    "t-checkbox-group": "tdesign-miniprogram/checkbox-group/checkbox-group",
-    "t-swipe-cell": "tdesign-miniprogram/swipe-cell/swipe-cell",
-    "t-sticky": "tdesign-miniprogram/sticky/sticky",
-    "t-fab": "tdesign-miniprogram/fab/fab",
-    "t-tab-bar": "tdesign-miniprogram/tab-bar/tab-bar",
-    "t-tab-bar-item": "tdesign-miniprogram/tab-bar/tab-bar-item",
-    "t-transition": "tdesign-miniprogram/transition/transition",
-    "t-indexes": "tdesign-miniprogram/indexes/indexes",
-    "t-dropdown-menu": "tdesign-miniprogram/dropdown-menu/dropdown-menu",
-    "t-dropdown-item": "tdesign-miniprogram/dropdown-menu/dropdown-item",
-    "t-drawer": "tdesign-miniprogram/drawer/drawer",
-    "t-pull-down-refresh": "tdesign-miniprogram/pull-down-refresh/pull-down-refresh",
-    "t-footer": "tdesign-miniprogram/footer/footer",
-    "t-divider": "tdesign-miniprogram/divider/divider",
-    "t-empty": "tdesign-miniprogram/empty/empty",
-    "t-back-top": "tdesign-miniprogram/back-top/back-top",
-    "t-grid-item": "tdesign-miniprogram/grid/grid-item",
-    "t-grid": "tdesign-miniprogram/grid/grid",
-    "t-count-down": "tdesign-miniprogram/count-down/count-down",
-    "t-image": "tdesign-miniprogram/image/image",
-    "t-search": "tdesign-miniprogram/search/search",
-    "t-navbar": "tdesign-miniprogram/navbar/navbar",
-    "t-date-time-picker": "tdesign-miniprogram/date-time-picker/date-time-picker",
-    "t-action-sheet": "tdesign-miniprogram/action-sheet/action-sheet",
-    "t-image-viewer": "tdesign-miniprogram/image-viewer/image-viewer",
-    "t-link": "tdesign-miniprogram/link/link"
+    "t-demo": "./components/demo-block/index"
   },
   "window": {
     "backgroundTextStyle": "light",
