+++ src/collapse-panel/collapse-panel.json
