+++ src/collapse-panel/collapse-panel.less
