+++ src/collapse-panel/collapse-panel.wxml
