+++ src/collapse-panel/props.ts
@@ -18,12 +18,12 @@ const props: TdCollapsePanelProps = {
   /** 禁止当前面板展开，优先级大于 Collapse 的同名属性 */
   disabled: {
     type: Boolean,
-    value: null,
+    value: undefined,
   },
   /** 当前折叠面板展开图标，优先级大于 Collapse 的同名属性 */
   expandIcon: {
     type: Boolean,
-    value: true,
+    value: undefined,
   },
   /** 组件类名，用于组件外层元素、标题、内容 */
   externalClasses: {
