+++ src/collapse/_example/placement/index.json
@@ -2,6 +2,6 @@
   "component": true,
   "usingComponents": {
     "t-collapse": "tdesign-miniprogram/collapse/collapse",
-    "t-collapse-panel": "tdesign-miniprogram/collapse/collapse-panel"
+    "t-collapse-panel": "tdesign-miniprogram/collapse-panel/collapse-panel"
   }
 }
