+++ src/collapse/index.ts
@@ -2,7 +2,5 @@ export { default as Collapse } from './collapse';
 
 export * from './type';
 export * from './props';
-export * from './collapse-panel-props';
 
 export { CollapseProps } from './collapse';
-export { CollapsePanelProps } from './collapse-panel';
